# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Cleaning Services API
  description: API para gestao de prestadores de servico - recibos, orcamentos e NFS-e
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor local
components:
  schemas:
...
    ReceiptResponse:
      type: object
      description: Dados de um recibo salvo no banco
      properties:
        id:
          type: integer
          example: 1
        receiptNumber:
          type: integer
          example: 5
        clientName:
          type: string
          example: Vivo VMT
        value:
          type: number
          example: 80.00
        emitterId:
          type: integer
          example: 1
        date:
          type: string
          format: date
          example: "2026-06-11"
    QuoteItemResponse:
      type: object
      properties:
        description:
          type: string
          example: "Limpeza de vitrine"
        value:
          type: number
          example: 928.99
        quantity:
          type: number
          example: 1.0
        unit:
          type: string
          example: un
        total:
          type: number
          example: 928.99
    QuoteResponse:
      type: object
      description: Dados de um orçamento salvo no banco
      properties:
        id:
          type: integer
          example: 1
        quoteNumber:
          type: integer
          example: 4
        emitterId:
          type: integer
          example: 2
        clientName:
          type: string
          example: JM Moveis Design
        message:
          type: string
          example: "Aqui esta o orcamento conforme solicitado"
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuoteItemResponse'
        discountPercent:
          type: number
          example: 25.0
        discountAmount:
          type: number
          example: 232.25
        subtotal:
          type: number
          example: 928.99
        finalValue:
          type: number
          example: 696.74
        paymentConditions:
          type: string
          example: Mensal
        paymentMethods:
          type: string
          example: "PIX, Dinheiro, Cartao de Debito, Boleto"
        date:
          type: string
          format: date
          example: "2026-06-11"
    ReceiptRequest:
      type: object
      description: Dados para gerar um recibo PDF
      required: [receiptNumber, clientName, value, emitterId]
      properties:
        receiptNumber:
          type: integer
          example: 1
        clientName:
          type: string
          example: Vivo VMT
        value:
          type: number
          example: 80.00
        emitterId:
          type: integer
          example: 1
    QuoteItemRequest:
      type: object
      description: Um item do orcamento - representa uma linha da tabela
      required: [description, value, quantity]
      properties:
        description:
          type: string
          example: "Limpeza de vitrine. Fachada parte alta R$529,99."
        value:
          type: number
          example: 928.99
        quantity:
          type: number
          example: 1.0
        unit:
          type: string
          default: un
          example: un
    QuoteRequest:
      type: object
      description: Dados para gerar o PDF do orcamento
      required: [emitterId, quoteNumber, clientName, items]
      properties:
        emitterId:
          type: integer
          example: 1
        quoteNumber:
          type: integer
          example: 4
        clientName:
          type: string
          example: JM Moveis Design
        message:
          type: string
          example: "Aqui esta o orcamento conforme solicitado"
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuoteItemRequest'
        discountPercent:
          type: number
          example: 25.0
        paymentConditions:
          type: string
          example: Mensal
        paymentMethods:
          type: string
          example: "PIX, Dinheiro, Cartao de Debito, Boleto"
...
paths:
...
  /receipts:
    post:
      summary: Gerar recibo PDF
      description: Gera o PDF do recibo e salva o registro no banco
      tags: [Receipts]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReceiptRequest'
      responses:
        '200':
          description: PDF gerado com sucesso
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Dados invalidos
        '404':
          description: Emitente nao encontrado
    get:
      summary: Listar todos os recibos gerados
      tags: [Receipts]
      responses:
        '200':
          description: Lista de recibos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReceiptResponse'
  /receipts/{id}:
    get:
      summary: Buscar recibo por ID
      tags: [Receipts]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Recibo encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
        '404':
          description: Recibo nao encontrado
  /quotes:
    post:
      summary: Gerar orcamento PDF
      description: Gera PDF com cabecalho do emitente, tabela de itens, desconto e condicoes de pagamento.
      tags: [Quotes]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QuoteRequest'
      responses:
        '200':
          description: PDF do orcamento gerado com sucesso
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '400':
          description: Dados invalidos
        '404':
          description: Emitente nao encontrado
    get:
      summary: Listar todos os orcamentos gerados
      tags: [Quotes]
      responses:
        '200':
          description: Lista de orcamentos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/QuoteResponse'
  /quotes/{id}:
    get:
      summary: Buscar orcamento por ID
      tags: [Quotes]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Orcamento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuoteResponse'
        '404':
          description: Orcamento nao encontrado
...